[tools/memory] Remove DWARF scanning in pa_tcache_inspect.
Now that we scan chrome's memory, there is no need to keep DWARF parsing
around. Besides simplifying the code, the DWARF path is subject to
bitrot, as it is not compiled on bots, and requires extra local
dependencies to work.

Remove it entirely.

Bug: 1169157
Change-Id: Ib62849a3a52bdd283acada140bcca4ece49dfe0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3171059
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#922911}
NOKEYCHECK=True
GitOrigin-RevId: d333862925cb5583aedfaf0e3bb428184dc2a92b